Finalized Factory Console

+++ FactoryWarehouseSystem/FactoryWarehouseSystem/IssueTON.cs
@@ -14,7 +14,9 @@ namespace FactoryWarehouseSystem
     public partial class IssueTON : Form
     {
         DataTable dt, dt1, dt2;
-
+        TransferOutNote ton;
+        Database db,db1;
+       
         public IssueTON()
         {
             InitializeComponent();
@@ -22,16 +24,15 @@ namespace FactoryWarehouseSystem
 
         private void btnClear_Click(object sender, EventArgs e)
         {
-            cmbItemCode.Text = "";
-            cmbItemName.Text = "";
+            cmbItemCode.Text = "";            
             txtQty.Clear();
             txtUnitPrice.Clear();
         }
 
         private void btnAddItem_Click(object sender, EventArgs e)
         {
-            Database db = new Database();
-            if (string.IsNullOrEmpty(cmbItemCode.Text) || string.IsNullOrEmpty(cmbItemName.Text) || string.IsNullOrEmpty(txtQty.Text) || string.IsNullOrEmpty(txtUnitPrice.Text))
+            db = new Database();
+            if (string.IsNullOrEmpty(cmbItemCode.Text) || string.IsNullOrEmpty(txtItemName.Text) || string.IsNullOrEmpty(txtQty.Text) || string.IsNullOrEmpty(txtUnitPrice.Text))
             {
                 MessageBox.Show("Empty Fields!");
             }
@@ -45,7 +46,7 @@ namespace FactoryWarehouseSystem
                 int mark = 0;
                 foreach (DataRow row in dt.Rows)
                 {
-                    if (row.Field<int>(1)==Convert.ToInt32(cmbItemCode.Text))
+                    if (row.Field<int>(1)==Convert.ToInt32(cmbItemCode.SelectedValue.ToString()))
                     {
                         mark = 1;
                         break;
@@ -54,13 +55,13 @@ namespace FactoryWarehouseSystem
 
                 if (mark==1)
                 {
-                    string updateQuery = "update tonTemp set qty = qty+" + txtQty.Text + " where itemID = " + cmbItemCode.Text + "";
+                    string updateQuery = "update tonTemp set qty = qty+" + txtQty.Text + " where itemID = " + cmbItemCode.SelectedValue.ToString() + "";
                     //MessageBox.Show(updateQuery);
                     db.inserUpdateDelete(updateQuery);
                 }
                 else
                 {
-                    string queryTemp = " insert into tonTemp values (" + txtTONNo.Text + " ," + cmbItemCode.Text + ",'" + cmbItemName.Text + "'," + txtQty.Text + "," + txtUnitPrice.Text + ")";
+                    string queryTemp = " insert into tonTemp values (" + txtTONNo.Text + " ," + cmbItemCode.SelectedValue.ToString() + ",'" + txtItemName.Text + "'," + txtQty.Text + "," + txtUnitPrice.Text + ")";
                     //MessageBox.Show(queryTemp);
                     db.inserUpdateDelete(queryTemp);
                 }
@@ -73,40 +74,40 @@ namespace FactoryWarehouseSystem
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            int selectedTONId = Convert.ToInt32(dataGridView1.CurrentRow.Cells[1].Value);
-            DialogResult dr = MessageBox.Show("Are you sure want to delete?", "Warning!", MessageBoxButtons.YesNo);
-            if (dr == DialogResult.Yes)
+            if (dataGridView1.CurrentCell!=null)
             {
-                Database db = new Database();
-                db.inserUpdateDelete("delete from tonTemp where itemID = " + selectedTONId + "");
-                dt = new DataTable();
-                dt = db.select("select * from tonTemp");
-                dataGridView1.DataSource = dt;
+                int selectedTONId = Convert.ToInt32(dataGridView1.CurrentRow.Cells[1].Value);
+                DialogResult dr = MessageBox.Show("Are you sure want to delete?", "Warning!", MessageBoxButtons.YesNo);
+                if (dr == DialogResult.Yes)
+                {
+                    Database db = new Database();
+                    db.inserUpdateDelete("delete from tonTemp where itemID = " + selectedTONId + "");
+                    dt = new DataTable();
+                    dt = db.select("select * from tonTemp");
+                    dataGridView1.DataSource = dt;
+                }
             }
         }
 
         private void cmbItemCode_SelectedIndexChanged(object sender, EventArgs e)
         {
-            Database db = new Database();
-            string code = db.getValue("select itemName from item where itemCode = '" + cmbItemCode.SelectedText + "'");
-            cmbItemName.Text = code;
+            db = new Database();
+            dt = new DataTable();
+            dt = db.select("select itemName, unitPrice from item where itemCode = '" + cmbItemCode.Text + "'");
+            txtUnitPrice.Text = dt.Rows[0][1].ToString();
+            txtItemName.Text = dt.Rows[0][0].ToString();
         }
 
         private void btnIssue_Click(object sender, EventArgs e)
         {
-            TransferOutNote ton = new TransferOutNote();
+            ton = new TransferOutNote();
             ton.FromLocation = txtFromLocation.Text;
             ton.Destination = cmbDestination.Text;
             ton.Id = Convert.ToInt32(txtTONNo.Text);
             ton.addTON();           
             MessageBox.Show("TON added successfully!");
             this.Close();
-        }
-
-        private void groupBox1_Enter(object sender, EventArgs e)
-        {
-
-        }
+        }       
 
         private void btnBack_Click(object sender, EventArgs e)
         {
@@ -115,34 +116,36 @@ namespace FactoryWarehouseSystem
 
         private void IssueTON_Load(object sender, EventArgs e)
         {
-            TransferOutNote ton = new TransferOutNote();
+            ton = new TransferOutNote();
             //Get Date
             txtDate.Text = DateTime.Now.ToString();
-            Database db = new Database();
+            db = new Database();
             db.inserUpdateDelete("delete from tonTemp");
             //Get TON ID
             string id;
             id = ton.getMaxId();
-            if (string.IsNullOrEmpty(id))
-            {
-                txtTONNo.Text = "1";
-            }
-            else
-            {
+
+            if (string.IsNullOrEmpty(id))            
+                txtTONNo.Text = "1";            
+            else            
                 txtTONNo.Text = (Convert.ToInt32(id) + 1).ToString();
-            }            
-            Database db1 = new Database();
+            
+            
             //Fill Combo 
             Item item = new Item();
-            dt1 = item.getItemDetails();            
+            dt1 = new DataTable();
+            dt1 = item.getItemDetails();                       
+            
+            cmbItemCode.DisplayMember = "code";
+            cmbItemCode.ValueMember = "ID";
             cmbItemCode.DataSource = dt1;
-            cmbItemCode.ValueMember = "itemID";
-            cmbItemName.DataSource = dt1;
-            cmbItemName.ValueMember = "itemName";
 
+            dt2 = new DataTable();
+            db1 = new Database();
             dt2 = db1.select("select * from warehouses");
-            cmbDestination.DataSource = dt2;
+            cmbDestination.DisplayMember = "location";
             cmbDestination.ValueMember = "location";
+            cmbDestination.DataSource = dt2;
         }
     }
 }
